+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 - Compatibility with Visual Studio 2019 (#448)
 - the pd_shift operation can now insert a shipment at the end of a route (#462)
+- Truncated distance value for end step (#463)
+- Zero distance value for last break in open-ended route (#475)
 
 ## [v1.9.0] - 2021-03-04